+++ README.md
@@ -2,7 +2,7 @@
  * @Author: ph4nt0mer
  * @Date: 2022-09-01 18:39:52
  * @LastEditors: rootphantomer
- * @LastEditTime: 2022-09-16 10:38:13
+ * @LastEditTime: 2022-09-19 10:45:40
  * @FilePath: /quake_go/README.md
  * @Description:
  *
@@ -15,7 +15,7 @@ Quake Command-Line Application With Golang
 
 ## rust 版本
 
-rust 项目代码 移步这里——https://github.com/360quake/quake_rs/
+rust 项目代码 移步这里——<https://github.com/360quake/quake_rs/>
 
 ## 安装
 
@@ -28,6 +28,11 @@ rust 项目代码 移步这里——https://github.com/360quake/quake_rs/
    ```
 
 ## 更新日志
+
+- 2022-09-14 v2.0:
+
+  - 支持其他项目可以引入 quake 模块
+  - import "quake/packages"
 
 - 2022-09-14 v1.7:
 
@@ -100,6 +105,31 @@ options:
 
 3、 ./quake search port:8080 -st 0 -sz 10
 
+```
+
+![alt](./iShot_2022-09-07_16.48.13.jpg)
+
+## 模块使用
+
+```golang
+import 360quake/quake/packages
+
+var reqjson packages.Reqjson
+reqjson.Query = hflag.GetString("args")
+reqjson.Start = hflag.GetString("start")
+reqjson.Size = hflag.GetString("size")
+reqjson.Start_time = hflag.GetTime("start_time")
+reqjson.End_time = hflag.GetTime("end_time")
+reqjson.Ignore_cache = hflag.GetBool("ignore_cache")
+reqjson.Field = hflag.GetString("field")
+reqjson.Query_txt = hflag.GetString("file_txt")
+
+
+packages.InfoGet(token.Token)
+packages.SearchServicePost(reqjson, token.Token)
+packages.HostSearchPost(reqjson, token.Token)
+
+返回包解析需要自己解析
 ```
 
 ![alt](./iShot_2022-09-07_16.48.13.jpg)
