+++ utils/setting.go
@@ -1,14 +1,14 @@
 /*
  * @Author: ph4nt0mer
  * @Date: 2022-09-01 15:44:48
- * @LastEditors: ph4nt0mer
- * @LastEditTime: 2022-09-01 15:44:50
- * @FilePath: /quake_go/setting.go
+ * @LastEditors: rootphantomer
+ * @LastEditTime: 2022-09-19 10:52:06
+ * @FilePath: /quake_go/utils/setting.go
  * @Description:
  *
  * Copyright (c) 2022 by ph4nt0mer, All Rights Reserved.
  */
-package setting
+package utils
 
 const HOST string = "https://quake.360.cn/api/"
 const VERSION string = "v3"
